% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMMixCure.R
\name{print.REsmcure}
\alias{print.REsmcure}
\title{Print REsmcure Object}
\usage{
\method{print}{REsmcure}(fit)
}
\arguments{
\item{fit}{an object of REsmcure.}
}
\description{
Output of REsmcure object.
}
\keyword{output}
