% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMixEM.R
\name{REem}
\alias{REem}
\title{Expectation-Maximization Algorithm (internal function)}
\usage{
REem(Time, Status, X, Z, coefs, varmat, grp, emmax, eps, idx)
}
\arguments{
\item{Time}{A vector of length n, containing the observed event/censoring time to event of interest.}

\item{Status}{A vector of length n, containing the event/censoring indicators (1=event, and 0=censoring).}

\item{X}{A nxk matrix of covariates for the latency model.}

\item{Z}{A nxk matrix of covariates for the incidence model.}

\item{coefs}{A stacked vector of length 2k+2j+1, corresponding to gamma, beta, random and frailty, as provided by initialisation stage.}

\item{varmat}{A 2x2 variance-covariance matrix of the bivariate gaussian random effects distribution.}

\item{grp}{A vector of length n, containing the grouping indicators (to identify the frailty term for each subject).}

\item{emmax}{specifieks the maximum number of iterations for the Expectation-Maximization algorithm. If the convergence criterion is not met, the EM iterations will be stopped after emmax iterations and the estimates will be based on the last maximum likelihood iteration. The default emmas = 50.}

\item{eps}{sets the convergence criterion (tolerance). If the summed squared changes in the parameters between iterations is lower than the specified value, the algorithm is considered to be converged.}

\item{idx}{Auxiliary parameter containing the identification indices of the sets of parameters.}
}
\description{
(Internal function) Implementation and Parameter optimization using the EM-algorithm.
}
\keyword{baseline}
\keyword{survival}
