% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REsmcure.R
\name{REsmcure}
\alias{REsmcure}
\title{Mixture Cure Model with Correlated Random Effects}
\usage{
REsmcure(formula, cureform, RE = "grp", rho = F, data, emmax = 50, eps = 1e-07)
}
\arguments{
\item{formula}{Double-sided formula ... .}

\item{cureform}{Right-hand-side formula for the latent incidence model.}

\item{RE}{String input identifying the variable name corresponding to the grouping variable.}

\item{rho}{Logical value. TRUE indicates estimation of the correlation parameter. FALSE sets the correlation to zero.}

\item{data}{data.frame for the dataset which contains all input data.}

\item{emmax}{Maximum number of iterations for the Expectation-Maximization Algorithm (default = 50).}

\item{eps}{Hyperparameter: convergence criterion (tolerance) on squared changes between iterations (default =1e-07).}
}
\description{
Implementation and Parameter optimization using the EM-algorithm.
}
\keyword{baseline}
\keyword{survival}
