% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REsmsurv.R
\name{REsmsurv}
\alias{REsmsurv}
\title{Estimate the baseline survival function S_0(t)}
\usage{
REsmsurv(Time, Status, X, beta, frailty, grp, w)
}
\arguments{
\item{Time}{A vector of length n, containing the observed event/censoring times.}

\item{Status}{A vector of length n, containing the event/censoring indicators (event:1 and censoring:0).}

\item{X}{A nxk matrix of covariates..}

\item{beta}{A vector of length k, containing the covariate effects.}

\item{frailty}{A vector of length j, containing all j frailty predictions.}

\item{grp}{A vector of length n, containing the grouping indicators (to identify the frailty term for each subject).}

\item{w}{A vector of length n, containing the expected susceptibility indicator.}
}
\description{
This function estimates the baseline survival function given the survival data, covariates and frailty predictions.
}
\keyword{baseline}
\keyword{survival}
