% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sebootjack.R
\name{sebootjack}
\alias{sebootjack}
\title{Simulate (Robust) Standard Errors using a bootstrap procedure or jackknife approach.}
\usage{
sebootjack(
  fit,
  data,
  method = c("boot", "jack"),
  k = 1,
  nboot = 100,
  nmax = 1000
)
}
\arguments{
\item{fit}{A fitted model of class smcure.}

\item{data}{The dataset on which the fitted model was fit.}

\item{method}{Either 'boot' for a bootstrap procedure, or 'jack' for a jackknife procedure.}

\item{k}{the number of observations to omit from the dataset for the leave-k-out Jackknife estimate on the standard errors. Default = 1, for al leave-one-out jackknife procedure.}

\item{nboot}{Number of iterations in the sampling-resampling procedure of the bootstrap estimator.}

\item{nmax}{Maximum number of resampling iterations for the bootstrap and jackknife estimators.}
}
\description{
Calculate (Robust) Standard Errors using either a bootstrap procedure or a Jackknife estimate.
}
\keyword{error}
\keyword{robust}
\keyword{standard}
