% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simsomdata.R
\name{simsomdata}
\alias{simsomdata}
\title{Simulate some datasets}
\usage{
simsomdata(i = 1, n, ng, coeffic, REvcov)
}
\arguments{
\item{i}{Seed value, integer (used for reproducible simulations and multi-dataset simulation).}

\item{n}{Sample size n, integer.}

\item{ng}{Number of higher-level units (groups), integer.}

\item{REvcov}{A list of size 3, containing the covariance-matrix elements (varcure, frailty and correlation parameter).}

\item{coeff}{A list of size 2, containing a vector of parameter values each, respectively of the indicence and the latency model.}
}
\description{
This function simulates some survival data with a cure proportion.
}
\keyword{simulate}
